refactor: use toast to display error message in add and detail pages
Refactor the code in the add and detail pages of the muzakki and mustahik components to use the toast component for displaying error messages instead of logging them to the console. This improves the user experience by providing a more visible and user-friendly way of presenting errors. The toast component is used to show the error message with a destructive variant, indicating the severity of the error.

Related to recent commits:
- feat: add NavMasterData component to AppSidebar
- feat: create mustahik CRUD pages and functionality
- feat: add delete functionality in muzakki
- feat: create detail muzakki page